requirements.txt
caldav>=1.3.0
icalendar>=5.0.0
notion-client>=2.2.0
python-dotenv>=1.0.0
typer>=0.9.0
